+++ packages/profire-flamegraph/CHANGELOG.md
@@ -3,7 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.0.2](https://github.com/TLin-Y/pyroscope/compare/@pyroscope/flamegraph@0.35.1...@profire/flamegraph@0.0.1) (2025-04-26)
+
+## [0.0.3]([Release 0.0.3 · TLin-Y/pyroscope](https://github.com/TLin-Y/pyroscope/releases/tag/profire-flamegraph)) (2025-04-28)
+
+Features
+
+* Avoid displaying useless unit string < 0.01
+* Fix GraphViz view bugs
+* Allow control of unit levels (from 1 to 6: ns, us, ms, sec, min, hour) via flamebear metadata 'unitLevel'
+* By default, convert samples (ns) to seconds by using samples(ns) / 1e9 (sample rate)
+
+## [0.0.2]([Release 0.0.2 · TLin-Y/pyroscope](https://github.com/TLin-Y/pyroscope/releases/tag/profire-flamegraph)) (2025-04-26)
 
 Features